apps/workers/pyproject.toml
[project]
name = "peptideiq-workers"
version = "0.1.0"
description = "PeptideIQ research engine - scraper workers and idea extraction pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "httpx>=0.28.0",
    "praw>=7.8.0",
    "pytrends>=4.9.0",
    "playwright>=1.49.0",
    "beautifulsoup4>=4.13.0",
    "feedparser>=6.0.0",
    "redis>=5.2.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "anthropic>=0.42.0",
    "openai>=1.58.0",
    "numpy>=2.0.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.0",
]
